Split out time system calls in the libc backend.
