+++ src/imp/libc/syscalls.rs
@@ -3,6 +3,7 @@
 pub(crate) use super::fs::syscalls::*;
 pub(crate) use super::net::syscalls::*;
 pub(crate) use super::process::syscalls::*;
+pub(crate) use super::time::syscalls::*;
 
 #[cfg(any(target_os = "ios", target_os = "macos"))]
 use super::conv::nonnegative_ret;
@@ -18,6 +19,8 @@ use super::io::MremapFlags;
 #[cfg(not(any(target_os = "ios", target_os = "macos", target_os = "wasi")))]
 use super::io::PipeFlags;
 use super::io::PollFd;
+#[cfg(not(target_os = "wasi"))]
+use super::io::{DupFlags, MapFlags, MprotectFlags, ProtFlags, Termios, Winsize};
 #[cfg(any(target_os = "android", target_os = "linux"))]
 use super::io::{EventfdFlags, UserfaultfdFlags};
 #[cfg(any(target_os = "android", target_os = "linux"))]
@@ -25,11 +28,12 @@ use super::io::{MlockFlags, ReadWriteFlags};
 #[cfg(not(target_os = "wasi"))]
 use super::offset::libc_mmap;
 use super::offset::{libc_pread, libc_pwrite};
+#[cfg(not(target_os = "redox"))]
+use super::offset::{libc_preadv, libc_pwritev};
 #[cfg(all(target_os = "linux", target_env = "gnu"))]
 use super::offset::{libc_preadv2, libc_pwritev2};
 #[cfg(target_os = "linux")]
 use super::rand::GetRandomFlags;
-use super::time::Timespec;
 use crate::io::{self, OwnedFd, RawFd};
 #[cfg(not(target_os = "wasi"))]
 use crate::process::Pid;
@@ -43,20 +47,8 @@ use std::ffi::CStr;
 use std::ffi::CString;
 use std::io::{IoSlice, IoSliceMut};
 use std::mem::MaybeUninit;
-#[cfg(not(any(target_os = "redox", target_os = "wasi",)))]
-use std::ptr::null_mut;
 #[cfg(not(any(target_os = "redox", target_env = "newlib")))]
 use std::sync::atomic::{AtomicUsize, Ordering};
-#[cfg(not(target_os = "wasi"))]
-use {
-    super::io::{DupFlags, MapFlags, MprotectFlags, ProtFlags, Termios, Winsize},
-    super::time::{ClockId, DynamicClockId},
-};
-#[cfg(not(target_os = "redox"))]
-use {
-    super::offset::{libc_preadv, libc_pwritev},
-    crate::time::NanosleepRelativeResult,
-};
 
 pub(crate) fn read(fd: BorrowedFd<'_>, buf: &mut [u8]) -> io::Result<usize> {
     let nread = unsafe {
@@ -725,134 +717,6 @@ pub(crate) fn getrandom(buf: &mut [u8], flags: GetRandomFlags) -> io::Result<usi
         ))?
     };
     Ok(nread as usize)
-}
-
-#[cfg(not(any(target_os = "redox", target_os = "wasi")))]
-#[inline]
-#[must_use]
-pub(crate) fn clock_getres(id: ClockId) -> Timespec {
-    let mut timespec = MaybeUninit::<Timespec>::uninit();
-    unsafe {
-        let _ = libc::clock_getres(id as libc::clockid_t, timespec.as_mut_ptr());
-        timespec.assume_init()
-    }
-}
-
-#[cfg(not(target_os = "wasi"))]
-#[inline]
-#[must_use]
-pub(crate) fn clock_gettime(id: ClockId) -> Timespec {
-    let mut timespec = MaybeUninit::<Timespec>::uninit();
-    // Use `unwrap()` here because `clock_getres` can fail if the clock itself
-    // overflows a number of seconds, but if that happens, the monotonic clocks
-    // can't maintain their invariants, or the realtime clocks aren't properly
-    // configured.
-    unsafe {
-        ret(libc::clock_gettime(
-            id as libc::clockid_t,
-            timespec.as_mut_ptr(),
-        ))
-        .unwrap();
-        timespec.assume_init()
-    }
-}
-
-#[cfg(not(target_os = "wasi"))]
-#[inline]
-pub(crate) fn clock_gettime_dynamic(id: DynamicClockId) -> io::Result<Timespec> {
-    let mut timespec = MaybeUninit::<Timespec>::uninit();
-    unsafe {
-        let id: libc::clockid_t = match id {
-            DynamicClockId::Known(id) => id as libc::clockid_t,
-
-            #[cfg(any(target_os = "android", target_os = "linux"))]
-            DynamicClockId::Dynamic(fd) => {
-                use io::AsRawFd;
-                const CLOCKFD: i32 = 3;
-                (!fd.as_raw_fd() << 3) | CLOCKFD
-            }
-
-            #[cfg(not(any(target_os = "android", target_os = "linux")))]
-            DynamicClockId::Dynamic(_fd) => {
-                // Dynamic clocks are not supported on this platform.
-                return Err(io::Error::INVAL);
-            }
-
-            #[cfg(any(target_os = "android", target_os = "linux"))]
-            DynamicClockId::RealtimeAlarm => libc::CLOCK_REALTIME_ALARM,
-
-            #[cfg(any(target_os = "android", target_os = "linux"))]
-            DynamicClockId::Tai => libc::CLOCK_TAI,
-
-            #[cfg(any(target_os = "android", target_os = "linux"))]
-            DynamicClockId::Boottime => libc::CLOCK_BOOTTIME,
-
-            #[cfg(any(target_os = "android", target_os = "linux"))]
-            DynamicClockId::BoottimeAlarm => libc::CLOCK_BOOTTIME_ALARM,
-        };
-
-        ret(libc::clock_gettime(
-            id as libc::clockid_t,
-            timespec.as_mut_ptr(),
-        ))?;
-
-        Ok(timespec.assume_init())
-    }
-}
-
-#[cfg(not(any(
-    target_os = "emscripten",
-    target_os = "freebsd", // FreeBSD 12 has clock_nanosleep, but libc targets FreeBSD 11.
-    target_os = "ios",
-    target_os = "macos",
-    target_os = "openbsd",
-    target_os = "redox",
-    target_os = "wasi",
-)))]
-#[inline]
-pub(crate) fn clock_nanosleep_relative(id: ClockId, request: &Timespec) -> NanosleepRelativeResult {
-    let mut remain = MaybeUninit::<Timespec>::uninit();
-    let flags = 0;
-    unsafe {
-        match libc::clock_nanosleep(id as libc::clockid_t, flags, request, remain.as_mut_ptr()) {
-            0 => NanosleepRelativeResult::Ok,
-            err if err == io::Error::INTR.0 => {
-                NanosleepRelativeResult::Interrupted(remain.assume_init())
-            }
-            err => NanosleepRelativeResult::Err(io::Error(err)),
-        }
-    }
-}
-
-#[cfg(not(any(
-    target_os = "freebsd", // FreeBSD 12 has clock_nanosleep, but libc targets FreeBSD 11.
-    target_os = "emscripten",
-    target_os = "ios",
-    target_os = "macos",
-    target_os = "openbsd",
-    target_os = "redox",
-    target_os = "wasi",
-)))]
-#[inline]
-pub(crate) fn clock_nanosleep_absolute(id: ClockId, request: &Timespec) -> io::Result<()> {
-    let flags = libc::TIMER_ABSTIME;
-    match unsafe { libc::clock_nanosleep(id as libc::clockid_t, flags, request, null_mut()) } {
-        0 => Ok(()),
-        err => Err(io::Error(err)),
-    }
-}
-
-#[cfg(not(target_os = "redox"))]
-#[inline]
-pub(crate) fn nanosleep(request: &Timespec) -> NanosleepRelativeResult {
-    let mut remain = MaybeUninit::<Timespec>::uninit();
-    unsafe {
-        match ret(libc::nanosleep(request, remain.as_mut_ptr())) {
-            Ok(()) => NanosleepRelativeResult::Ok,
-            Err(io::Error::INTR) => NanosleepRelativeResult::Interrupted(remain.assume_init()),
-            Err(err) => NanosleepRelativeResult::Err(err),
-        }
-    }
 }
 
 #[cfg(any(target_os = "android", target_os = "linux"))]
